+++ bin/check-supervisor.rb
@@ -26,6 +26,7 @@
 
 require 'sensu-plugin/check/cli'
 require 'ruby-supervisor'
+require 'set'
 
 class CheckSupervisor < Sensu::Plugin::Check::CLI
   option :host,
@@ -84,9 +85,12 @@ class CheckSupervisor < Sensu::Plugin::Check::CLI
       critical "Tried to access #{config[:host]} but failed: #{e}"
     end
 
+    failed_processes = Set.new
     @super.processes.each do |process|
-      critical "#{process['name']} not running: #{process['statename'].downcase}" if config[:critical].include?(process['statename'])
+      failed_processes << "#{process['name']} not running: #{process['statename'].downcase}" if config[:critical].include?(process['statename'])
     end
+
+    critical "\n#{failed_processes.to_a.join("\n")}" unless failed_processes.empty?
 
     ok 'All processes running'
   end # def run